+++ src/cow-react/modules/common/pure/CurrencyArrowSeparator/index.cosmos.tsx
@@ -1,4 +1,7 @@
-import { CurrencyArrowSeparator, CurrencyArrowSeparatorProps } from 'cow-react/common/pure/CurrencyArrowSeparator'
+import {
+  CurrencyArrowSeparator,
+  CurrencyArrowSeparatorProps,
+} from 'cow-react/modules/common/pure/CurrencyArrowSeparator'
 
 const defaultProps: CurrencyArrowSeparatorProps = {
   isLoading: false,
