+++ src/cow-react/modules/common/pure/CurrencyArrowSeparator/index.tsx
