+++ src/cow-react/modules/common/pure/CurrencyArrowSeparator/style-mixins.tsx
