+++ src/cow-react/modules/common/pure/CurrencyArrowSeparator/styled.tsx
