+++ src/cow-react/modules/common/pure/CurrencyInputPanel/styled.tsx
