+++ src/cow-react/modules/common/pure/CurrencyInputPanel/typings.ts
